fix: broken syntax on configuration pages (#6)
Because of the nature of `vim.inspect` and the generation of default
values in configuration documentation there are cases where the
following would be added:

```
something = {
  ["key"] = <function 1>
}
```

This fix prevents this by replacing these function tags with actual Lua
code. In the future this should probably read the functions from source
as is done in other cases.