feat: improve code formatting for default values in configurations (#7)
The implementation for showing referenced function from the original
docgen wasn't really working well. This has been removed and now will
show a function with a comment to redirect to the appropriate
documentation or source code.

The lua code is now also run through stylua in order to get consistent
syntax.